api/requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.0
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-flex-fields==0.3.5
drf-nested-routers==0.91
furl==2.0.0
gunicorn==19.9.0
inflection==0.3.1
orderedmultidict==1.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
